+++ CHANGELOG.md
@@ -29,7 +29,7 @@
 - Manufactoring Order: add rollback in operation order as it is in manufacturing order when click on finish
 - Web service mobile: change the "create timesheet line" method to update it also, and have duration/hours updated
 - Product : displaying quantities in stock with big numbers (until millions)
-
+- Manufacturing order: fix NPE when adding manually an operation order, caused by missing work center
 
 ## [5.0.1] - 2018-06-18
 ## Improvements